+++ WalletSystem/Transfer.cs
@@ -4,6 +4,6 @@
     {
         public Wallet from { get; private set; }
         public Wallet to { get; private set; }
-        public Transfer( double tempSum, Wallet f, Wallet t) : base( tempSum) { from = f; to = t; }
+        public Transfer( decimal tempSum, Wallet f, Wallet t) : base( tempSum) { from = f; to = t; }
     }
 }
